Add provider label to promxy
