+++ service/controller/resource/resource.go
@@ -164,6 +164,7 @@ func New(config Config) ([]resource.Interface, error) {
 			K8sClient:    config.K8sClient,
 			Logger:       config.Logger,
 			Installation: config.Installation,
+			Provider:     config.Provider,
 		}
 
 		promxyResource, err = promxy.New(c)
